+++ src/modules/mysql/MySQLPoolConnection.ts
@@ -1,5 +1,5 @@
 import { FieldPacket, Pool, createPool } from "mysql2/promise"
-import { DatabaseObjectDelete, DatabaseObjectInsert, DatabaseObjectUpdate, OrderByConfig, PoolConnectionConfig, QueryConfigDelete, QueryConfigFetch, QueryConfigInsert, QueryConfigUpdate, QueryResultField, QueryResultFieldFlag, QueryResultRow, SQLPoolConnection, Timer } from "../core"
+import { DatabaseObjectDelete, DatabaseObjectInsert, DatabaseObjectUpdate, OrderByConfig, PoolConnectionConfig, QueryConfigBase, QueryConfigDelete, QueryConfigFetch, QueryConfigInsert, QueryConfigUpdate, QueryResultField, QueryResultFieldFlag, QueryResultRow, SQLPoolConnection, Timer, buildSQLQueryDeletions, buildSQLQueryInsertions, buildSQLQueryUpdates } from "../core"
 import { mysqlTypeMap } from "./types"
 
 export class MySQLPoolConnection implements SQLPoolConnection {
@@ -132,7 +132,7 @@ export class MySQLPoolConnection implements SQLPoolConnection {
         const orderBy = this.createOrderBy(config.orderBy)
         const where = config.filterString ? 'WHERE ' + config.filterString : ''
         const query = `SELECT *
-FROM ${config.database}.${config.table}
+FROM ${this.dbId(config)}
 ${where}
 ${orderBy}
 ${limit}`
@@ -140,43 +140,37 @@ ${limit}`
     }
 
     buildQueriesInsert(insertions: DatabaseObjectInsert, queryConfig: QueryConfigInsert) {
-        const queries: string[] = []
-        for (let i = 0; i < insertions.insertions.length; i++) {
-            const insertion = insertions.insertions[i]
-            const queryStart = `INSERT INTO ${queryConfig.database}.${queryConfig.table}`
-            const queryFields = ` (${Object.keys(insertion).join(', ')})`
-            const queryValues = ` VALUES (${Object.keys(insertion).map(fieldName => this.convertJavascriptValueToSQL(insertion[fieldName]))})`
-            queries.push(queryStart + queryFields + queryValues)
-        }
-        return queries
+        return buildSQLQueryInsertions(
+            insertions,
+            {
+                dbId: this.dbId(queryConfig),
+                stringWrap: '"'
+            }
+        )
     }
 
     buildQueriesDelete(deletions: DatabaseObjectDelete[], queryConfig: QueryConfigDelete) {
-        const queries = []
-        for (let i = 0; i < deletions.length; i++) {
-            const whereStatements = "WHERE " + Object.keys(deletions[i].where).map(field =>
-                `${field} = ${this.convertJavascriptValueToSQL(deletions[i].where[field])}`
-            ).join(' AND ')
-            queries.push(`DELETE FROM ${queryConfig.database}.${queryConfig.table}
-${whereStatements}`)
-        }
-        return queries
+        return buildSQLQueryDeletions(
+            deletions,
+            {
+                dbId: this.dbId(queryConfig),
+                stringWrap: '"'
+            }
+        )
     }
 
-    buildQueriesUpdate(changes: DatabaseObjectUpdate[], config: QueryConfigUpdate) {
-        const queries = []
-        for (let i = 0; i < changes.length; i++) {
-            const whereStatements = "WHERE " + Object.keys(changes[i].where).map(field =>
-                `${field} = ${this.convertJavascriptValueToSQL(changes[i].where[field])}`
-            ).join(' AND ')
-            const setStatements = "SET " + Object.keys(changes[i].update).map(field =>
-                `${field} = ${this.convertJavascriptValueToSQL(changes[i].update[field])}`
-            ).join(',')
-            queries.push(`UPDATE ${config.database}.${config.table}
-${setStatements}
-${whereStatements}`)
-        }
-        return queries
+    buildQueriesUpdate(changes: DatabaseObjectUpdate[], queryConfig: QueryConfigUpdate) {
+        return buildSQLQueryUpdates(
+            changes,
+            {
+                dbId: this.dbId(queryConfig),
+                stringWrap: '"'
+            }
+        )
+    }
+
+    private dbId(queryConfig: QueryConfigBase) {
+        return `${queryConfig.database}.${queryConfig.table}`
     }
 
     async executeQuery(query: string) {
@@ -267,16 +261,6 @@ ${whereStatements}`)
             flags.push('notnull')
         }
         return flags
-    }
-
-    private convertJavascriptValueToSQL(val: string | null | undefined | number) {
-        if (val === null) {
-            return 'NULL'
-        }
-        if (typeof val === 'number') {
-            return val
-        }
-        return `"${val}"`
     }
 
     private async query(query: string) {